refactor: Rename ConductingController to OverallController in Conducting namespace
