+++ app/Http/Controllers/Project/Conducting/OverallController.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Http\Controllers\Project;
+namespace App\Http\Controllers\Project\Conducting;
 
 use App\Http\Controllers\Controller;
 use App\Models\Project;
 use Illuminate\Http\Request;
 
-class ConductingController extends Controller
+class OverallController extends Controller
 {
 
     public function index(string $id_project) {
@@ -14,4 +14,6 @@ class ConductingController extends Controller
         $project = Project::findOrFail($id_project);
         return view('project.conducting.index', compact('project'));
     }
+
+    
 }
